+++ forest/shared/src/cli/config.rs
@@ -7,7 +7,7 @@ use std::{path::PathBuf, sync::Arc};
 use forest_chain_sync::SyncConfig;
 use forest_db::db_engine::DbConfig;
 use forest_libp2p::Libp2pConfig;
-use forest_networks::ChainConfig;
+use forest_networks::{ChainConfig, NetworkChain};
 use log::LevelFilter;
 use serde::{Deserialize, Serialize};
 use url::Url;
@@ -70,22 +70,33 @@ pub struct SnapshotFetchConfig {
 #[derive(Serialize, Deserialize, PartialEq, Eq)]
 pub struct FilecoinSnapshotFetchConfig {
     pub mainnet: Url,
+    pub mainnet_compressed: Url,
     pub calibnet: Url,
+    pub calibnet_compressed: Url,
 }
 
 impl Default for FilecoinSnapshotFetchConfig {
     fn default() -> Self {
         // unfallible unwrap as we know that the value is correct
+        // <https://lotus.filecoin.io/lotus/manage/chain-management/#lightweight-snapshot>
         Self {
             /// Default `mainnet` snapshot URL. The assumption is that it will
             /// redirect once and will contain a `sha256sum` file
             /// with the same URL (but different extension).
             mainnet: Url::try_from("https://snapshots.mainnet.filops.net/minimal/latest").unwrap(),
+            mainnet_compressed: Url::try_from(
+                "https://snapshots.mainnet.filops.net/minimal/latest.zst",
+            )
+            .unwrap(),
             /// Default `calibnet` snapshot URL. The assumption is that it will
             /// redirect once and will contain a `sha256sum` file
             /// with the same URL (but different extension).
             calibnet: Url::try_from("https://snapshots.calibrationnet.filops.net/minimal/latest")
                 .unwrap(),
+            calibnet_compressed: Url::try_from(
+                "https://snapshots.calibrationnet.filops.net/minimal/latest.zst",
+            )
+            .unwrap(),
         }
     }
 }
@@ -184,6 +195,18 @@ pub struct Config {
     pub log: LogConfig,
     pub snapshot_fetch: SnapshotFetchConfig,
     pub tokio: TokioConfig,
+}
+
+impl Config {
+    pub fn from_chain(network_type: &NetworkChain) -> Self {
+        match network_type {
+            NetworkChain::Mainnet => Config::default(),
+            NetworkChain::Calibnet => Config {
+                chain: Arc::new(ChainConfig::calibnet()),
+                ..Config::default()
+            },
+        }
+    }
 }
 
 impl Config {