+++ node/forest_libp2p/src/behaviour.rs
@@ -3,10 +3,8 @@
 
 use ahash::{HashMap, HashSet};
 use forest_libp2p_bitswap::BitswapBehaviour;
-use forest_utils::encoding::blake2b_256;
+use forest_utils::{encoding::blake2b_256, version::FOREST_VERSION_STRING};
 use libp2p::{
-    allow_block_list::{self, BlockedPeers},
-    connection_limits::{self, ConnectionLimits},
     core::identity::Keypair,
     gossipsub::{
         self, IdentTopic as Topic, MessageAuthenticity, MessageId, PublishError, SubscriptionError,
@@ -38,8 +36,6 @@ pub(crate) struct ForestBehaviour {
     discovery: DiscoveryBehaviour,
     ping: ping::Behaviour,
     identify: identify::Behaviour,
-    connection_limits: connection_limits::Behaviour,
-    pub(super) blocked_peers: allow_block_list::Behaviour<BlockedPeers>,
     pub(super) hello: HelloBehaviour,
     pub(super) chain_exchange: ChainExchangeBehaviour,
     pub(super) bitswap: BitswapBehaviour,
@@ -57,12 +53,7 @@ impl Recorder<ForestBehaviourEvent> for Metrics {
 }
 
 impl ForestBehaviour {
-    pub fn new(
-        local_key: &Keypair,
-        config: &Libp2pConfig,
-        network_name: &str,
-        connection_limits: ConnectionLimits,
-    ) -> Self {
+    pub fn new(local_key: &Keypair, config: &Libp2pConfig, network_name: &str) -> Self {
         let mut gs_config_builder = gossipsub::ConfigBuilder::default();
         gs_config_builder.max_transmit_size(1 << 20);
         gs_config_builder.validation_mode(ValidationMode::Strict);
@@ -105,16 +96,15 @@ impl ForestBehaviour {
             .with_user_defined(config.bootstrap_peers.clone())
             .target_peer_count(config.target_peer_count as u64);
 
+        warn!("libp2p Forest version: {}", FOREST_VERSION_STRING.as_str());
         ForestBehaviour {
             gossipsub,
             discovery: discovery_config.finish(),
             ping: Default::default(),
-            identify: identify::Behaviour::new(identify::Config::new(
-                "ipfs/0.1.0".into(),
-                local_key.public(),
-            )),
-            connection_limits: connection_limits::Behaviour::new(connection_limits),
-            blocked_peers: Default::default(),
+            identify: identify::Behaviour::new(
+                identify::Config::new("ipfs/0.1.0".into(), local_key.public())
+                    .with_agent_version(format!("forest-{}", FOREST_VERSION_STRING.as_str())),
+            ),
             bitswap,
             hello: HelloBehaviour::default(),
             chain_exchange: ChainExchangeBehaviour::default(),