+++ utils/forest_shim/src/machine.rs
@@ -1,0 +1,6 @@
+// Copyright 2019-2023 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+pub use fvm::machine::Manifest as ManifestV2;
+mod manifest;
+pub use manifest::Manifest;