+++ CHANGELOG.md
@@ -33,6 +33,48 @@
 
 ### Fixed
 
+- [#2796] (https://github.com/ChainSafe/forest/pull/2796): Fix issue when
+  running Forest on calibnet using a configuration file only.
+- [#2807] (https://github.com/ChainSafe/forest/pull/2807): Fix issue with v11
+  actor CIDs.
+- [#2804] (https://github.com/ChainSafe/forest/pull/2804): Add work around for
+  FVM bug that caused `forest-cli sync wait` to fail.
+
+## Forest v0.8.1 "Cold Exposure"
+
+### Fixed
+
+- [#2788](https://github.com/ChainSafe/forest/pull/2788): Move back to the
+  upstream `ref-fvm` and bump the dependency version so that it included the
+  latest critical [patch](https://github.com/filecoin-project/ref-fvm/pull/1750)
+
+## Forest v0.8.0 "Jungle Speed" (2023-04-21)
+
+### Added
+
+- [#2763](https://github.com/ChainSafe/forest/issues/2763): Support NV19 and
+  NV20. ⛈️
+
+## Forest v0.7.2 "Roberto" (2023-04-19)
+
+### Added
+
+- [#2741](https://github.com/ChainSafe/forest/issues/2741): Support importing
+  zstd compressed snapshot car files
+- [#2741](https://github.com/ChainSafe/forest/issues/2741): Support fetching
+  zstd compressed snapshots with filecoin provider via `--compressed` option
+- [#1454](https://github.com/ChainSafe/forest/issues/1454): Added state
+  migration support for NV18.
+
+### Changed
+
+- [#2770](https://github.com/ChainSafe/forest/issues/2767): Use `latest` tag for
+  stable releases, and `edge` for latest development builds.
+
+### Removed
+
+### Fixed
+
 ## Forest v0.7.1 (2023-03-29)
 
 Notable updates: